T-008: PR-2: Move section_panel.tsx to flyout_v2 + update legacy
